Cambio de Javascript a Typescript
