+++ api/src/models/Address.ts
@@ -1,19 +1,30 @@
-const { DataTypes } = require('sequelize');
+import { DataTypes } from 'sequelize';
 // Exportamos una funcion que define el modelo
 // Luego le injectamos la conexion a sequelize.
-module.exports = (sequelize) => {
+module.exports = (sequelize: any) => {
   // defino el modelo
-  sequelize.define('category', {
+  sequelize.define('address', {
     id: {
       type: DataTypes.INTEGER,
       autoIncrement: true,
       allowNull: false,
       primaryKey: true
     },
-    category: {
-      type: DataTypes.STRING,
+    id_user: {
+      type: DataTypes.INTEGER,
       allowNull: false,
-      unique: true
+      references: {
+        model: 'users',
+        key: 'id'
+      }
+    },
+    address: {
+      type: DataTypes.STRING,
+      allowNull: false
+    },
+    zip_code: {
+      type: DataTypes.INTEGER,
+      allowNull: false,
     },
     isActive: {
       allowNull: false,
