+++ api/src/models/Carrousel.ts
@@ -1,7 +1,7 @@
-const { DataTypes } = require('sequelize');
+import { DataTypes } from 'sequelize';
 // Exportamos una funcion que define el modelo
 // Luego le injectamos la conexion a sequelize.
-module.exports = (sequelize) => {
+module.exports = (sequelize: any) => {
   // defino el modelo
   sequelize.define('carrousel', {
     id: {
