+++ api/src/models/Category.ts
@@ -1,16 +1,16 @@
-const { DataTypes } = require('sequelize');
+import { DataTypes } from 'sequelize';
 // Exportamos una funcion que define el modelo
 // Luego le injectamos la conexion a sequelize.
-module.exports = (sequelize) => {
+module.exports = (sequelize: any) => {
   // defino el modelo
-  sequelize.define('size', {
+  sequelize.define('category', {
     id: {
       type: DataTypes.INTEGER,
       autoIncrement: true,
       allowNull: false,
       primaryKey: true
     },
-    size: {
+    category: {
       type: DataTypes.STRING,
       allowNull: false,
       unique: true
