+++ api/src/models/Color.ts
@@ -1,13 +1,13 @@
 const { DataTypes } = require('sequelize');
 // Exportamos una funcion que define el modelo
 // Luego le injectamos la conexion a sequelize.
-module.exports = (sequelize) => {
+module.exports = (sequelize: any) => {
   // defino el modelo
-  sequelize.define('color',{
-    id:{
+  sequelize.define('color', {
+    id: {
       type: DataTypes.INTEGER,
       autoIncrement: true,
-      allowNull:false,
+      allowNull: false,
       primaryKey: true
     },
     color: {
@@ -20,5 +20,5 @@ module.exports = (sequelize) => {
       type: DataTypes.BOOLEAN,
       defaultValue: true
     }
-  },{timestamps:false, freezeTableName: true});
+  }, { timestamps: false, freezeTableName: true });
 };
