+++ api/src/models/Images.ts
@@ -1,30 +1,19 @@
-const { DataTypes } = require('sequelize');
+import { DataTypes } from 'sequelize';
 // Exportamos una funcion que define el modelo
 // Luego le injectamos la conexion a sequelize.
-module.exports = (sequelize) => {
+module.exports = (sequelize: any) => {
   // defino el modelo
-  sequelize.define('address', {
+  sequelize.define('images', {
     id: {
       type: DataTypes.INTEGER,
       autoIncrement: true,
       allowNull: false,
       primaryKey: true
     },
-    id_user: {
-      type: DataTypes.INTEGER,
-      allowNull: false,
-      references: {
-        model: 'users',
-        key: 'id'
-      }
-    },
-    address: {
+    image: {
       type: DataTypes.STRING,
-      allowNull: false
-    },
-    zip_code: {
-      type: DataTypes.INTEGER,
       allowNull: false,
+      unique: true
     },
     isActive: {
       allowNull: false,
