+++ api/src/models/Orders.ts
@@ -1,8 +1,8 @@
-const { DataTypes } = require('sequelize');
+import { DataTypes } from 'sequelize';
 // Exportamos una funcion que define el modelo
 // Luego le injectamos la conexion a sequelize.
 
-module.exports = (sequelize) => {
+module.exports = (sequelize: any) => {
     // defino el modelo
     sequelize.define('orders', {
         id: {
