+++ api/src/models/Size.ts
@@ -1,16 +1,17 @@
-const { DataTypes } = require('sequelize');
+import { DataTypes } from 'sequelize';
 // Exportamos una funcion que define el modelo
 // Luego le injectamos la conexion a sequelize.
-module.exports = (sequelize) => {
+
+module.exports = (sequelize: any) => {
   // defino el modelo
-  sequelize.define('images', {
+  sequelize.define('size', {
     id: {
       type: DataTypes.INTEGER,
       autoIncrement: true,
       allowNull: false,
       primaryKey: true
     },
-    image: {
+    size: {
       type: DataTypes.STRING,
       allowNull: false,
       unique: true
