+++ api/src/routes/ejemplo ruta.ts
@@ -1,0 +1,13 @@
+'use strict'
+// se requiere el models
+import { Router } from 'express';
+
+const router = Router();
+router.get('/', async (_req, _res) => {
+
+})
+router.post('/', async (_req, _res) => {
+
+})
+
+module.exports = router;
